re-jigger the save state
